requirements.txt
virtualenv
streamlit
google-generativeai
python-dotenv
streamlit_modal
